+++ react-app/src/components/Profile/index.js
@@ -1,0 +1,10 @@
+import UserInfoData from './UserInfoData';
+import UserInfoForm from './UserInfoForm';
+
+export default function UserInfo({ showUpdateUser, setShowUpdateUser }) {
+  if (showUpdateUser) {
+    return <UserInfoForm setShowUpdateUser={setShowUpdateUser} />;
+  } else {
+    return <UserInfoData setShowUpdateUser={setShowUpdateUser} />;
+  }
+}
